requirements.txt
asgiref==3.10.0
colorama==0.4.6
crispy-bootstrap5==2024.10
Django==5.0.14
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.1.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
